updated codes

+++ api/authors/delete.php
@@ -1,15 +1,11 @@
-
 <?php
-  // Headers
-  header('Access-Control-Allow-Origin: *');
-  header('Content-Type: application/json');
+// Headers
   header('Access-Control-Allow-Methods: DELETE');
-  header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow-Methods, Authorization,X-Requested-With');
-
-
+  header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow- Methods, Authorization,X-Requested-With');
 
   include_once '../../config/Database.php';
   include_once '../../models/Author.php';
+
   // Instantiate DB & connect
   $database = new Database();
   $db = $database->connect();
@@ -20,16 +16,29 @@
   // Get raw data
   $data = json_decode(file_get_contents("php://input"));
 
+  // Create Author, if...
+  if (!isset($inputdata['author'])) {
+    echo json_encode(['message' => 'Missing Required Parameters']);
+    exit;
+  }
+
+  // Get raw data
+   $data = json_decode(file_get_contents("php://input"));
+
   // Set ID to UPDATE
   $author->id = $data->id;
 
-  // Delete author
+  $data = json_decode(file_get_contents("php://input"), true);
+
+  //  // if delete auth ...
   if($author->delete()) {
     echo json_encode(
-      array('message' => 'Author deleted')
+      array('id' => $data['id'])
     );
+	//else...
   } else {
     echo json_encode(
       array('message' => 'Author not deleted')
     );
   }
+?>
